Moved the CRC code from a header to a compiled .c file.
As the XC8 compiler does not support inline functions
in its evaluation version, it was useless to declare the crc function
as 'inline' and define it in a header file.

Modified the CRC table generator to declare the crc_table as 'extern
const' instead of 'const' as it implies internal linkage (i.e. the
variable is not accessible at linker scope when compiling the whole
project).

Adapted the unit tests according to those modifications.